+++ piecrust/src/lib.rs
@@ -125,6 +125,7 @@ pub use call_tree::{CallTree, CallTreeElem};
 pub use contract::{ContractData, ContractDataBuilder};
 pub use error::Error;
 pub use session::{CallReceipt, Session, SessionData};
+pub use store::PageOpening;
 pub use vm::{HostQuery, VM};
 
 // re-export the contents of the `piecrust-uplink` crate wholesale, ensuring