+++ piecrust/src/store.rs
@@ -9,25 +9,28 @@
 mod bytecode;
 mod memory;
 mod metadata;
-mod objectcode;
+mod module;
 mod session;
 mod tree;
 
 use std::collections::btree_map::Entry::*;
 use std::collections::{BTreeMap, BTreeSet};
+use std::fmt::{Debug, Formatter};
 use std::path::{Path, PathBuf};
 use std::sync::mpsc;
 use std::{fs, io, thread};
 
-pub use bytecode::Bytecode;
-use dusk_wasmtime::{Engine, Module};
-pub use memory::{Memory, MAX_MEM_SIZE, PAGE_SIZE};
-pub use metadata::Metadata;
-pub use objectcode::Objectcode;
+use dusk_wasmtime::Engine;
 use piecrust_uplink::ContractId;
 use session::ContractDataEntry;
-pub use session::ContractSession;
 use tree::{ContractIndex, Hash};
+
+pub use bytecode::Bytecode;
+pub use memory::{Memory, PAGE_SIZE};
+pub use metadata::Metadata;
+pub use module::Module;
+pub use session::ContractSession;
+pub use tree::PageOpening;
 
 const BYTECODE_DIR: &str = "bytecode";
 const MEMORY_DIR: &str = "memory";
@@ -36,11 +39,22 @@ const OBJECTCODE_EXTENSION: &str = "a";
 const METADATA_EXTENSION: &str = "m";
 
 /// A store for all contract commits.
-#[derive(Debug)]
 pub struct ContractStore {
     sync_loop: thread::JoinHandle<()>,
+    engine: Engine,
+
     call: mpsc::Sender<Call>,
     root_dir: PathBuf,
+}
+
+impl Debug for ContractStore {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("ContractStore")
+            .field("sync_loop", &self.sync_loop)
+            .field("call", &self.call)
+            .field("root_dir", &self.root_dir)
+            .finish()
+    }
 }
 
 impl ContractStore {
@@ -53,13 +67,13 @@ impl ContractStore {
     /// [`commit`]: ContractSession::commit
     /// [`delete`]: ContractStore::delete_commit
     /// [`session spawning`]: ContractStore::session
-    pub fn new<P: AsRef<Path>>(engine: &Engine, dir: P) -> io::Result<Self> {
+    pub fn new<P: AsRef<Path>>(engine: Engine, dir: P) -> io::Result<Self> {
         let root_dir = dir.as_ref();
 
         fs::create_dir_all(root_dir)?;
 
         let (call, calls) = mpsc::channel();
-        let commits = read_all_commits(engine, root_dir)?;
+        let commits = read_all_commits(&engine, root_dir)?;
 
         let loop_root_dir = root_dir.to_path_buf();
 
@@ -71,6 +85,7 @@ impl ContractStore {
 
         Ok(Self {
             sync_loop,
+            engine,
             call,
             root_dir: root_dir.into(),
         })
@@ -144,7 +159,12 @@ impl ContractStore {
     }
 
     fn session_with_base(&self, base: Option<Commit>) -> ContractSession {
-        ContractSession::new(&self.root_dir, base, self.call.clone())
+        ContractSession::new(
+            &self.root_dir,
+            self.engine.clone(),
+            base,
+            self.call.clone(),
+        )
     }
 }
 
@@ -205,32 +225,24 @@ fn commit_from_dir<P: AsRef<Path>>(
             ));
         }
 
-        let objectcode_path =
-            bytecode_path.with_extension(OBJECTCODE_EXTENSION);
+        let module_path = bytecode_path.with_extension(OBJECTCODE_EXTENSION);
 
-        if !objectcode_path.is_file() {
+        if !module_path.is_file() {
             return Err(io::Error::new(
                 io::ErrorKind::InvalidData,
-                format!("Non-existing objectcode for contract: {contract_hex}"),
+                format!("Non-existing module for contract: {contract_hex}"),
             ));
         }
 
         // SAFETY it is safe to deserialize the file here, since we don't use
         // the module here. We just want to check if the file is valid.
-        if unsafe {
-            Module::deserialize_file(engine, &objectcode_path).is_err()
-        } {
+        if Module::from_file(engine, &module_path).is_err() {
             let bytecode = Bytecode::from_file(bytecode_path)?;
             let module =
                 Module::new(engine, bytecode.as_ref()).map_err(|err| {
                     io::Error::new(io::ErrorKind::InvalidData, err)
                 })?;
-            fs::write(
-                objectcode_path,
-                module.serialize().map_err(|err| {
-                    io::Error::new(io::ErrorKind::InvalidData, err)
-                })?,
-            )?;
+            fs::write(module_path, module.serialize())?;
         }
 
         let memory_dir = memory_dir.join(&contract_hex);
@@ -499,12 +511,11 @@ fn write_commit_inner<P: AsRef<Path>>(
         }
 
         let bytecode_path = directories.bytecode_dir.join(&contract_hex);
-        let objectcode_path =
-            bytecode_path.with_extension(OBJECTCODE_EXTENSION);
+        let module_path = bytecode_path.with_extension(OBJECTCODE_EXTENSION);
         let metadata_path = bytecode_path.with_extension(METADATA_EXTENSION);
 
         // If the contract already existed in the base commit, we hard link the
-        // bytecode, objectcode, and metadata files to avoid duplicating them,
+        // bytecode, module, and metadata files to avoid duplicating them,
         // otherwise we write them to disk.
         //
         // Also, if there is a base commit, we hard link the pages of the
@@ -512,7 +523,7 @@ fn write_commit_inner<P: AsRef<Path>>(
         if let Some(base) = &directories.base {
             if let Some(elem) = base.inner.index.get(contract) {
                 let base_bytecode_path = base.bytecode_dir.join(&contract_hex);
-                let base_objectcode_path =
+                let base_module_path =
                     base_bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let base_metadata_path =
                     base_bytecode_path.with_extension(METADATA_EXTENSION);
@@ -520,7 +531,7 @@ fn write_commit_inner<P: AsRef<Path>>(
                 let base_memory_dir = base.memory_dir.join(&contract_hex);
 
                 fs::hard_link(base_bytecode_path, bytecode_path)?;
-                fs::hard_link(base_objectcode_path, objectcode_path)?;
+                fs::hard_link(base_module_path, module_path)?;
                 fs::hard_link(base_metadata_path, metadata_path)?;
 
                 for page_index in &elem.page_indices {
@@ -537,7 +548,7 @@ fn write_commit_inner<P: AsRef<Path>>(
             }
         } else {
             fs::write(bytecode_path, &contract_data.bytecode)?;
-            fs::write(objectcode_path, &contract_data.objectcode)?;
+            fs::write(module_path, &contract_data.module.serialize())?;
             fs::write(metadata_path, &contract_data.metadata)?;
         }
     }
@@ -549,13 +560,13 @@ fn write_commit_inner<P: AsRef<Path>>(
 
                 let bytecode_path =
                     directories.bytecode_dir.join(&contract_hex);
-                let objectcode_path =
+                let module_path =
                     bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let metadata_path =
                     bytecode_path.with_extension(METADATA_EXTENSION);
 
                 let base_bytecode_path = base.bytecode_dir.join(&contract_hex);
-                let base_objectcode_path =
+                let base_module_path =
                     base_bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let base_metadata_path =
                     base_bytecode_path.with_extension(METADATA_EXTENSION);
@@ -566,7 +577,7 @@ fn write_commit_inner<P: AsRef<Path>>(
                 fs::create_dir_all(&memory_dir)?;
 
                 fs::hard_link(base_bytecode_path, bytecode_path)?;
-                fs::hard_link(base_objectcode_path, objectcode_path)?;
+                fs::hard_link(base_module_path, module_path)?;
                 fs::hard_link(base_metadata_path, metadata_path)?;
 
                 for page_index in &elem.page_indices {